use model isntead
